+++ lilya/middleware/trustedhost.py
@@ -11,22 +11,28 @@ ENFORCE_DOMAIN_WILDCARD = "Domain wildcard patterns must be as per example '*.ex
 
 
 class TrustedHostMiddleware(MiddlewareProtocol):
+    scope_flag_name: str = "host_is_trusted"
+
     def __init__(
         self,
         app: ASGIApp,
-        allowed_hosts: typing.Sequence[str] | None = None,
+        allowed_hosts: typing.Iterable[str] | None = None,
         www_redirect: bool = True,
+        block_untrusted_hosts: bool = True,
     ) -> None:
         """
         Middleware for enforcing trusted host headers in incoming requests.
 
         Args:
             app (ASGIApp): The ASGI application to wrap.
-            allowed_hosts (Optional[Sequence[str]]): List of allowed host patterns. Defaults to ["*"].
-            www_redirect (bool): Whether to redirect requests with missing 'www.' to 'www.' prefixed URLs.
+            allowed_hosts (Optional[Iterable[str]]): List of allowed host patterns. Defaults to ["*"].
+            www_redirect (bool): Whether to redirect requests with missing 'www.' to 'www.' prefixed URLs. Defaults to True.
+            block_untrusted_hosts (bool): Whether to block the request or just add  = True,
         """
         if allowed_hosts is None:
-            allowed_hosts = ["*"]
+            allowed_hosts = {"*"}
+        else:
+            allowed_hosts = set(allowed_hosts)
 
         for pattern in allowed_hosts:
             assert "*" not in pattern[1:], ENFORCE_DOMAIN_WILDCARD
@@ -34,9 +40,10 @@ class TrustedHostMiddleware(MiddlewareProtocol):
                 assert pattern.startswith("*."), ENFORCE_DOMAIN_WILDCARD
 
         self.app = app
-        self.allowed_hosts = list(allowed_hosts)
-        self.allow_any = "*" in allowed_hosts
+        self.allowed_hosts = allowed_hosts
+        self.allow_any = "*" in self.allowed_hosts
         self.www_redirect = www_redirect
+        self.block_untrusted_hosts = block_untrusted_hosts
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
         """
@@ -56,6 +63,7 @@ class TrustedHostMiddleware(MiddlewareProtocol):
         is_valid_host, found_www_redirect = self.validate_host(host)
 
         if is_valid_host:
+            scope[self.scope_flag_name] = True
             await self.app(scope, receive, send)
         else:
             await self.handle_invalid_host(scope, receive, send, found_www_redirect)
@@ -92,11 +100,14 @@ class TrustedHostMiddleware(MiddlewareProtocol):
             send (Send): ASGI send channel.
             found_www_redirect (bool): Whether 'www.' redirect was found.
         """
-        response: Response
+        response: Response | ASGIApp
+        scope[self.scope_flag_name] = False
         if found_www_redirect and self.www_redirect:
             url = URL.build_from_scope(scope=scope)
             redirect_url = url.replace(netloc="www." + url.netloc)
             response = RedirectResponse(url=str(redirect_url))
-        else:
+        elif self.block_untrusted_hosts:
             response = PlainText("Invalid host header", status_code=400)
+        else:
+            response = self.app
         await response(scope, receive, send)