+++ docs/en/docs/release-notes.md
@@ -16,6 +16,18 @@ hide:
 
 - When adding encoders dependencies with `infer_body` as true, it was not filtering the body params properly.
 
+## 0.16.2
+
+### Added
+
+- Add `TrustedReferrerMiddleware`.
+- Add `SessionFixingMiddleware`.
+- Add `ClientIPScopeOnlyMiddleware` (splitted from `ClientIPMiddleware`).
+
+### Changed
+
+- Add `block_untrusted_hosts` parameter to `TrustedHostMiddleware`.
+
 ## 0.16.1
 
 ### Added